use one sigv4 file and move all tests to source/test/ folder (#186)
* init use only one sigv4 file

* Update test_api_unit_test.py

* Update sigv4.py

* copy shared folder for deployment

* Update config.json

* Update config.json

* cp sigv4 to zip

* Update build-s3-dist.sh

* move all tests to source/test/ folder

* Update run-unit-tests.sh

* Pre-commit fix

* Update config.json

* Update config.json

* Update pull-request-workflow.yml

* Update pull-request-workflow.yml

* update python path

* unit tests go in /source/api/tests/unit and all integration tests go in /source/api/tests/integ

* temporary commit to test workflow

* Update run_test.sh

* temp commit

* Update pull-request-workflow.yml

* AMC_ENDPOINT_URL -> AMC_API_ENDPOINT

* echo github secrets

* Update pull-request-workflow.yml

* Update pull-request-workflow.yml

* Update pull-request-workflow.yml

* Update pull-request-workflow.yml

* Update run_test.sh

* use github env variables

* Update pull-request-workflow.yml

* Update pull-request-workflow.yml

* Update run_test.sh

* Update run_test.sh

* Update run_test.sh

* Use script to run integ test. instead of github && pre-commit

* Update pull-request-workflow.yml

* Update pull-request-workflow.yml
