+++ .github/workflows/pull-request-workflow.yml
@@ -104,10 +104,8 @@ jobs:
           python3.10 -m venv "$VENV"
           source "$VENV"/bin/activate
           ls -l
-          cd source
-          ./tests/run_test.sh --run_unit_test
-          cd api/tests
-          ./run_api_test.sh --run_unit_test
+          cd source/tests
+          ./run_test.sh --run_unit_test
       - name: Run Coverage Report
         run: |
           cd deployment
@@ -230,8 +228,15 @@ jobs:
         uses: actions/checkout@v2.3.4
         with:
           ref: ${{ github.sha }}
-      - name: Run API Tests
-        id: api_test
+      - name: Run Integration Tests
+        id: integ_test
+        env:
+          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+          AWS_REGION: ${{needs.build.outputs.REGION}}
+          AMC_API_ENDPOINT: ${{ secrets.AMC_API_ENDPOINT }}
+          TEST_DATA_UPLOAD_ACCOUNT_ID: ${{ secrets.DATA_UPLOAD_ACCOUNT_ID }}
+          DATA_BUCKET_NAME: ${{ secrets.DATA_BUCKET_NAME }}
         run: |
           STACK_NAME=${{needs.deploy.outputs.STACK_NAME}}
           echo "STACK_NAME=$STACK_NAME" >> $GITHUB_OUTPUT
@@ -239,55 +244,18 @@ jobs:
           echo "SOLUTION_NAME=$SOLUTION_NAME" >> $GITHUB_OUTPUT
           VERSION=${{needs.build.outputs.VERSION}}
           echo "VERSION=$VERSION" >> $GITHUB_OUTPUT
-          RESOURCE_ID=$(aws cloudformation describe-stack-resources --stack-name $STACK_NAME --logical-resource-id GlueStack --query "StackResources[0].PhysicalResourceId")
-          echo "RESOURCE_ID=$RESOURCE_ID" >> $GITHUB_OUTPUT
-          NESTED_STACK_ID=$(echo $RESOURCE_ID | cut -d "/" -f 2)
-          echo "NESTED_STACK_ID=$NESTED_STACK_ID" >> $GITHUB_OUTPUT
-          AMC_GLUE_JOB_NAME=$(aws cloudformation describe-stack-resources --stack-name $NESTED_STACK_ID --logical-resource-id "AmcGlueJob" --query "StackResources[0].PhysicalResourceId")
-          AMC_GLUE_JOB_NAME=$(echo "$AMC_GLUE_JOB_NAME" | tr -d '"')
-          ARTIFACT_BUCKET=$(aws cloudformation describe-stack-resources --stack-name $STACK_NAME --logical-resource-id "ArtifactBucket" --query "StackResources[0].PhysicalResourceId")
-          ARTIFACT_BUCKET=$(echo $ARTIFACT_BUCKET | tr -d '"')
-          echo "ARTIFACT_BUCKET=$ARTIFACT_BUCKET" >> $GITHUB_OUTPUT
-          ROLE_NAME=$(aws cloudformation describe-stack-resources --stack-name $STACK_NAME --logical-resource-id "AmcApiAccessRole" --query "StackResources[0].PhysicalResourceId")
-          ROLE_NAME=$(echo $ROLE_NAME | tr -d '"')
-          echo "ROLE_NAME=$ROLE_NAME" >> $GITHUB_OUTPUT
-          CURRENT_ASSUME_ROLE_POLICY1=$(aws iam get-role --role-name $ROLE_NAME --query "Role.AssumeRolePolicyDocument.Statement[0]")
-          CURRENT_ASSUME_ROLE_POLICY2=$(aws iam get-role --role-name $ROLE_NAME --query "Role.AssumeRolePolicyDocument.Statement[1]")
           TEST_ACCOUNT_ID=$(aws sts get-caller-identity --query Account)
           TEST_ACCOUNT_ID=$(echo $TEST_ACCOUNT_ID | tr -d '"')
-          export TEST_ACCOUNT_ID=$TEST_ACCOUNT_ID
           TEST_USER_ARN="arn:aws:iam::${TEST_ACCOUNT_ID}:user/githubBuilderBot"
-          TEST_USER_ASSUME_ROLE_POLICY='{"Effect":"Allow","Principal":{"AWS":"'"$TEST_USER_ARN"'"},"Action":"sts:AssumeRole"}'
-          echo "TEST_USER_ASSUME_ROLE_POLICY=$TEST_USER_ASSUME_ROLE_POLICY" >> $GITHUB_OUTPUT
-          TEST_IAM_USER_POLICY='{"Version":"2012-10-17","Statement":['$CURRENT_ASSUME_ROLE_POLICY1', '$CURRENT_ASSUME_ROLE_POLICY2', '$TEST_USER_ASSUME_ROLE_POLICY']}'
-          aws iam update-assume-role-policy --role-name $ROLE_NAME --policy-document "$TEST_IAM_USER_POLICY"
-          AMC_GLUE_JOB_ROLE_NAME=$(aws cloudformation describe-stack-resources --stack-name $NESTED_STACK_ID --logical-resource-id "AmcGlueJobRole" --query "StackResources[0].PhysicalResourceId")
-          AMC_GLUE_JOB_ROLE_NAME=$(echo $AMC_GLUE_JOB_ROLE_NAME | tr -d '"')
-          SYSTEM_TABLE_NAME=$(aws cloudformation describe-stack-resources --stack-name $STACK_NAME --logical-resource-id "SystemTable" --query "StackResources[0].PhysicalResourceId")
-          SYSTEM_TABLE_NAME=$(echo $SYSTEM_TABLE_NAME | tr -d '"')
-          echo "AMC_GLUE_JOB_ROLE_NAME=$AMC_GLUE_JOB_ROLE_NAME" >> $GITHUB_ENV
-          echo "AMC_GLUE_JOB_ROLE_NAME=$AMC_GLUE_JOB_ROLE_NAME" >> $GITHUB_OUTPUT
-          export AMC_GLUE_JOB_NAME=$AMC_GLUE_JOB_NAME
-          export AMC_API_ROLE_ARN="arn:aws:iam::${TEST_ACCOUNT_ID}:role/$ROLE_NAME"
-          export DATA_BUCKET_NAME=${{ secrets.DATA_BUCKET_NAME }}
-          export ARTIFACT_BUCKET=$ARTIFACT_BUCKET
-          export SYSTEM_TABLE_NAME=$SYSTEM_TABLE_NAME
-          export AMC_API_ENDPOINT=${{ secrets.AMC_API_ENDPOINT }}
-          export SOLUTION_NAME=$SOLUTION_NAME
-          export CUSTOMER_MANAGED_KEY=''
-          export AWS_XRAY_CONTEXT_MISSING='LOG_ERROR'
-          export botoConfig="{}"
-          export VERSION=$VERSION
-          export AWS_REGION="us-east-1"
-          export TEST_DATA_UPLOAD_ACCOUNT_ID=${{secrets.DATA_UPLOAD_ACCOUNT_ID}}
-          export AWS_XRAY_SDK_ENABLED=false
-          aws iam attach-role-policy --role-name $AMC_GLUE_JOB_ROLE_NAME --policy-arn "arn:aws:iam::aws:policy/AmazonS3FullAccess"
-          VENV=$(mktemp -d) && echo "$VENV"
-          python3.10 -m venv "$VENV"
-          source "$VENV"/bin/activate
           ls -l
-          pip3 install -r source/api/tests/requirements-test.txt
-          pytest source/api/tests/test_api_integration.py -vv
+          cd source/tests/
+          ./run_test.sh --run_integ_test \
+          --stack-name $STACK_NAME \
+          --test_user_arn $TEST_USER_ARN \
+          --version $VERSION \
+          --solution_name $SOLUTION_NAME \
+          --aws-xray-sdk-enabled false \
+          --boto-config {}
 
   cleanup:
     if: always()
@@ -308,10 +276,6 @@ jobs:
         run: |
           aws s3 rb s3://${{needs.build.outputs.DIST_OUTPUT_BUCKET}}-${{needs.build.outputs.REGION}} --force
           aws s3 rb s3://${{needs.build.outputs.TEMPLATE_OUTPUT_BUCKET}} --force
-      - name: Detach s3 role
-        if: needs.integration_test.result == 'success' || needs.integration_test.result == 'failure'
-        run: |
-          aws iam detach-role-policy --role-name ${{needs.integration_test.outputs.AMC_GLUE_JOB_ROLE_NAME}} --policy-arn "arn:aws:iam::aws:policy/AmazonS3FullAccess"
       - name: Remove stack
         if: needs.deploy.result == 'success'
         run: |
