+++ deployment/build-s3-dist.sh
@@ -268,6 +268,9 @@ fi
 # Otherwise, chalice will use the existing deployment package
 [ -e .chalice/deployments ] && rm -rf .chalice/deployments
 
+echo "cp -R $source_dir/share ./dist"
+cp -R "$source_dir/share" "./dist"
+
 echo "Running chalice..."
 chalice package --merge-template external_resources.json dist
 echo "Finished running chalice."
@@ -277,6 +280,10 @@ if [ $? -ne 0 ]; then
   echo "ERROR: Failed to build api template"
   exit 1
 fi
+cd dist
+echo "deployment.zip /share/* -x **/__pycache__/*"
+zip -r deployment.zip share/* -x "**/__pycache__/*"
+cd ..
 echo "cp ./dist/deployment.zip $regional_dist_dir-api.zip"
 cp ./dist/deployment.zip "$regional_dist_dir"/api.zip
 if [ $? -ne 0 ]; then
@@ -349,6 +356,8 @@ echo "Building AMC uploader function"
 cd "$source_dir/amc_uploader" || exit 1
 [ -e dist ] && rm -r dist
 mkdir -p dist
+echo "cp -R $source_dir/share ./dist"
+cp -R "$source_dir/share" "./dist"
 [ -e package ] && rm -r package
 mkdir -p package
 echo "preparing packages from requirements.txt"
@@ -363,6 +372,10 @@ zip -q -r9 ../dist/amc_uploader.zip .
 popd || exit 1
 zip -q -g ./dist/amc_uploader.zip ./amc_uploader.py
 zip -q -g ./dist/amc_uploader.zip ./lib/sigv4.py
+cd dist
+echo "amc_uploader.zip share/* -x **/__pycache__/*"
+zip -r amc_uploader.zip share/* -x "**/__pycache__/*"
+cd ..
 cp "./dist/amc_uploader.zip" "$regional_dist_dir/amc_uploader.zip"
 
 echo "------------------------------------------------------------------------------"
