+++ source/tests/unit_test/test_api.py
@@ -6,7 +6,7 @@
 #   * Unit test for project api endpoints and workflow.
 #
 # USAGE:
-#   ./run_api_test.sh --run_unit_test
+#   ./run_test.sh --run_unit_test --test-file-name test_api.py
 ###############################################################################
 
 import json
@@ -133,7 +133,7 @@ def test_get_data_columns(test_configs, get_amc_json, test_data):
 
 
 @mock_sts
-@patch("chalicelib.sigv4.requests.Session")
+@patch("chalicelib.sigv4.sigv4.requests.Session")
 def test_delete_dataset(mock_session_response, test_configs):
     mock_session_response.mount = MagicMock()
     mock_session_response.return_value.delete.return_value = MagicMock(
@@ -159,7 +159,255 @@ def test_delete_dataset(mock_session_response, test_configs):
 
 
 @mock_sts
-@patch("chalicelib.sigv4.requests.Session")
+@patch("chalicelib.sigv4.sigv4.requests.Session")
+def test_list_datasets(mock_session_response, test_configs):
+    mock_session_response.mount = MagicMock()
+    test_body = {
+        "destination_endpoint": test_configs["destination_endpoint"],
+    }
+    expected_response = {
+        "dataSets": [
+            {
+                "columns": [
+                    {
+                        "dataType": "STRING",
+                        "description": "The customer resolved id",
+                        "isMainEventTime": False,
+                        "isMainUserId": True,
+                        "isMainUserIdType": False,
+                        "isNullable": True,
+                        "name": "user_id",
+                        "requiresOneWayHashing": False,
+                    },
+                    {
+                        "dataType": "STRING",
+                        "description": "The customer resolved type",
+                        "isMainEventTime": False,
+                        "isMainUserId": False,
+                        "isMainUserIdType": True,
+                        "isNullable": True,
+                        "name": "user_type",
+                        "requiresOneWayHashing": False,
+                    },
+                    {
+                        "dataType": "STRING",
+                        "description": "hashed First name",
+                        "externalUserIdType": {
+                            "type": "HashedIdentifier",
+                            "identifierType": "FIRST_NAME",
+                        },
+                        "isMainEventTime": False,
+                        "isMainUserId": False,
+                        "isMainUserIdType": False,
+                        "isNullable": True,
+                        "name": "first_name",
+                        "requiresOneWayHashing": False,
+                    },
+                    {
+                        "dataType": "STRING",
+                        "description": "hashed Last name",
+                        "externalUserIdType": {
+                            "type": "HashedIdentifier",
+                            "identifierType": "LAST_NAME",
+                        },
+                        "isMainEventTime": False,
+                        "isMainUserId": False,
+                        "isMainUserIdType": False,
+                        "isNullable": True,
+                        "name": "last_name",
+                        "requiresOneWayHashing": False,
+                    },
+                    {
+                        "dataType": "TIMESTAMP",
+                        "description": "Timestamp",
+                        "isMainEventTime": True,
+                        "isMainUserId": False,
+                        "isMainUserIdType": False,
+                        "isNullable": False,
+                        "name": "timestamp",
+                        "requiresOneWayHashing": False,
+                    },
+                    {
+                        "columnType": "DIMENSION",
+                        "dataType": "STRING",
+                        "description": "Product quantity",
+                        "isMainEventTime": False,
+                        "isMainUserId": False,
+                        "isMainUserIdType": False,
+                        "isNullable": False,
+                        "name": "product_quantity",
+                        "requiresOneWayHashing": False,
+                    },
+                    {
+                        "columnType": "METRIC",
+                        "dataType": "STRING",
+                        "description": "Product name",
+                        "isMainEventTime": False,
+                        "isMainUserId": False,
+                        "isMainUserIdType": False,
+                        "isNullable": False,
+                        "name": "product_name",
+                        "requiresOneWayHashing": False,
+                    },
+                ],
+                "compressionFormat": "GZIP",
+                "createdTime": "2023-03-14T20:39:19.043Z",
+                "dataSetId": "iantest12",
+                "dataSetType": "FACT",
+                "fileFormat": "JSON",
+                "owner": "CUSTOMER",
+                "period": "P1D",
+                "updatedTime": "2023-03-14T20:39:19.043Z",
+            },
+        ]
+    }
+    mock_session_response.return_value.get.return_value = MagicMock(
+        status_code=200, text=json.dumps(expected_response)
+    )
+
+    content_type = test_configs["content_type"]
+    with Client(app.app) as client:
+        response = client.http.post(
+            "/list_datasets",
+            headers={"Content-Type": content_type},
+            body=json.dumps(test_body),
+        )
+        assert response.status_code == 200
+        assert response.json_body == expected_response
+
+
+@mock_sts
+@patch("chalicelib.sigv4.sigv4.requests.Session")
+def test_describe_dataset(mock_session_response, test_configs):
+    mock_session_response.mount = MagicMock()
+    test_body = {
+        "destination_endpoint": test_configs["destination_endpoint"],
+        "dataSetId": test_configs["data_set_id"],
+    }
+    expected_response = {
+        "columns": [
+            {
+                "dataType": "STRING",
+                "description": "The customer resolved id",
+                "isMainEventTime": False,
+                "isMainUserId": True,
+                "isMainUserIdType": False,
+                "isNullable": True,
+                "name": "user_id",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "dataType": "STRING",
+                "description": "The customer resolved type",
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": True,
+                "isNullable": True,
+                "name": "user_type",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "dataType": "STRING",
+                "description": "hashed First name",
+                "externalUserIdType": {
+                    "type": "HashedIdentifier",
+                    "identifierType": "FIRST_NAME",
+                },
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": False,
+                "isNullable": True,
+                "name": "first_name",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "dataType": "STRING",
+                "description": "hashed Last name",
+                "externalUserIdType": {
+                    "type": "HashedIdentifier",
+                    "identifierType": "LAST_NAME",
+                },
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": False,
+                "isNullable": True,
+                "name": "last_name",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "dataType": "STRING",
+                "description": "hashed Email",
+                "externalUserIdType": {
+                    "type": "HashedIdentifier",
+                    "identifierType": "EMAIL",
+                },
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": False,
+                "isNullable": True,
+                "name": "email",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "dataType": "STRING",
+                "description": "Product quantity",
+                "externalUserIdType": {"type": "LiveRamp"},
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": False,
+                "isNullable": False,
+                "name": "product_quantity",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "columnType": "DIMENSION",
+                "dataType": "STRING",
+                "description": "Timestamp",
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": False,
+                "isNullable": False,
+                "name": "timestamp",
+                "requiresOneWayHashing": False,
+            },
+            {
+                "columnType": "DIMENSION",
+                "dataType": "STRING",
+                "description": "Product name",
+                "isMainEventTime": False,
+                "isMainUserId": False,
+                "isMainUserIdType": False,
+                "isNullable": False,
+                "name": "product_name",
+                "requiresOneWayHashing": False,
+            },
+        ],
+        "compressionFormat": "GZIP",
+        "createdTime": "2023-03-22T19:26:35.034Z",
+        "dataSetId": "imgeo-normal",
+        "dataSetType": "DIMENSION",
+        "fileFormat": "JSON",
+        "owner": "CUSTOMER",
+        "period": "P1D",
+        "updatedTime": "2023-03-22T19:26:35.034Z",
+    }
+    mock_session_response.return_value.get.return_value = MagicMock(
+        status_code=200, text=json.dumps(expected_response)
+    )
+
+    content_type = test_configs["content_type"]
+    with Client(app.app) as client:
+        response = client.http.post(
+            "/describe_dataset",
+            headers={"Content-Type": content_type},
+            body=json.dumps(test_body),
+        )
+        assert response.status_code == 200
+        assert response.json_body == expected_response
+
+
+@mock_sts
+@patch("chalicelib.sigv4.sigv4.requests.Session")
 def test_upload_status(mock_session_response, test_configs):
     expected_data = {
         "sourceS3Bucket": "some_bucket",
@@ -239,7 +487,7 @@ def test_get_etl_jobs(test_configs):
 
 
 @mock_sts
-@patch("chalicelib.sigv4.requests.post")
+@patch("chalicelib.sigv4.sigv4.requests.post")
 def test_create_dataset(mock_response, test_configs):
     payload = {
         "destination_endpoint": test_configs["destination_endpoint"],
@@ -315,7 +563,7 @@ def test_start_amc_transformation(test_configs):
             assert glue_resp["JobRun"]["Id"] == response.json_body["JobRunId"]
 
 
-@patch("chalicelib.sigv4.requests.post")
+@patch("chalicelib.sigv4.sigv4.requests.post")
 def test_system_configuration(mock_response, test_configs):
     with mock_iam(), mock_s3(), mock_dynamodb():
         content_type = test_configs["content_type"]
